allow `--verbose` flag to globally set log level
