+++ tooling/cli/src/lib.rs
@@ -14,20 +14,12 @@ mod plugin;
 mod signer;
 
 use clap::{FromArgMatches, IntoApp, Parser, Subcommand};
-
 use std::ffi::OsString;
-
-pub(crate) trait CommandExt {
-  fn pipe(&mut self) -> Result<&mut Self>;
-}
-
-impl CommandExt for std::process::Command {
-  fn pipe(&mut self) -> Result<&mut Self> {
-    self.stdout(os_pipe::dup_stdout()?);
-    self.stderr(os_pipe::dup_stderr()?);
-    Ok(self)
-  }
-}
+use std::process::{Output, Command};
+use env_logger::fmt::Color;
+use env_logger::{Builder};
+use log::{log_enabled, Level, debug};
+use std::io::Write;
 
 #[derive(serde::Deserialize)]
 pub struct VersionMetadata {
@@ -49,7 +41,7 @@ pub struct VersionMetadata {
 )]
 struct Cli {
   /// Enables verbose logging
-  #[clap(short, long, parse(from_occurrences))]
+  #[clap(short, long, global = true, parse(from_occurrences))]
   verbose: usize,
   #[clap(subcommand)]
   command: Commands,
@@ -99,6 +91,38 @@ where
     Err(e) => e.exit(),
   };
 
+  let mut builder = Builder::from_default_env();
+  builder
+    .format_indent(Some(12))
+    .filter(None, level_from_usize(cli.verbose).to_level_filter())
+    .format(|f, record| {
+      if let Some(action) = record.key_values().get("action".into()) {
+        let mut action_style = f.style();
+        action_style.set_color(Color::Green).set_bold(true);
+
+        write!(f, "{:>12} ", action_style.value(action.to_str().unwrap()))?;
+      } else {
+        let mut level_style = f.default_level_style(record.level());
+        level_style.set_bold(true);
+
+        write!(
+          f,
+          "{:>12} ",
+          level_style.value(prettyprint_level(record.level()))
+        )?;
+      }
+
+      if log_enabled!(Level::Debug) {
+        let mut target_style = f.style();
+        target_style.set_color(Color::Black);
+
+        write!(f, "[{}] ", target_style.value(record.target()))?;
+      }
+
+      writeln!(f, "{}", record.args())
+    })
+    .init();
+
   match cli.command {
     Commands::Build(options) => build::command(options)?,
     Commands::Dev(options) => dev::command(options)?,
@@ -109,4 +133,46 @@ where
   }
 
   Ok(())
+}
+
+/// This maps the occurrence of `--verbose` flags to the correct log level 
+fn level_from_usize(num: usize) -> Level {
+  match num {
+    0 => Level::Info,
+    1 => Level::Debug,
+    2.. => Level::Trace,
+    _ => panic!()
+  }
+}
+
+/// The default string representation for `Level` is all uppercaps which doesn't mix well with the other printed actions.
+fn prettyprint_level(lvl: Level) -> &'static str {
+  match lvl {
+    Level::Error => "Error",
+    Level::Warn => "Warn",
+    Level::Info => "Info",
+    Level::Debug => "Debug",
+    Level::Trace => "Trace",
+  }
+}
+
+pub trait CommandExt {
+  fn output_ok(&mut self) -> crate::Result<Output>;
+}
+
+impl CommandExt for Command {
+  fn output_ok(&mut self) -> crate::Result<Output> {
+    debug!(action = "Running"; "Command `{} {}`", self.get_program().to_string_lossy(), self.get_args().map(|arg| arg.to_string_lossy()).fold(String::new(), |acc, arg| format!("{} {}", acc, arg)));
+
+    let output = self.output()?;
+
+    debug!("Stdout: {}", String::from_utf8_lossy(&output.stdout));
+    debug!("Stderr: {}", String::from_utf8_lossy(&output.stderr));
+
+    if output.status.success() {
+      Ok(output)
+    } else {
+      Err(anyhow::anyhow!(String::from_utf8_lossy(&output.stderr).to_string()))
+    }
+  }
 }
